requirements.txt
geoip2==4.2.0
IPy==1.1
uvicorn==0.13.4
fastapi==0.66.0
starlette==0.14.2
pydantic==1.8.2
ipip-ipdb
jinja2